refactor(config): extract resolveHandlerContext (shared createHandlerOptions prelude)
createHandlerOptions.{server,client} ran a byte-identical prelude before their
per-side work: merge defaults, resolve auto-discovered files, build the route URL,
load the route file paths (throw on failure). Extract it to resolveHandlerContext
in the shared module; each variant destructures the result.

Follows the worker-block dedup; covered by the existing createHandlerOptions
characterization test (it exercises this whole prelude).

Verified: build clean, 101 entrypoints, test:unit 472, test:server 655/4-skip.
Advances 66g.4 (the remaining server-only component-loader is genuinely per-side,
not a dedup opportunity).
